back/update ticket query api

+++ back/dto/comm.go
@@ -4,7 +4,7 @@ type JsonStruct struct {
 	Code    int         `json:"code"`
 	Message interface{} `json:"msg"`
 	Data    interface{} `json:"data"`
-	Count   int64       `json:"count"`
+	Count   int         `json:"count"`
 }
 
 type JsonErrorStruct struct {
